fix: 更新公网IP至8.148.236.60，修复Spring Security编译兼容问题
- 前端 request.js / product.js 的 baseURL 和 AI 识别接口指向新 IP
- DataInitializer storeId 数据类型 Long→Integer 修正
- SecurityConfig authorizeHttpRequests→authorizeRequests 适配 Spring Security 5.7.x

+++ 连锁超市微服务管理系统/supermarket-cloud/supermarket-auth/src/main/java/com/supermarket/auth/config/DataInitializer.java
@@ -21,13 +21,13 @@ public class DataInitializer implements CommandLineRunner {
     public void run(String... args) {
         String encodedPassword = passwordEncoder.encode("123456");
 
-        initUser("admin", encodedPassword, "HQ", "超级管理员", 0L);
-        initUser("store1", encodedPassword, "STORE", "一号门店店长", 1L);
-        initUser("store2", encodedPassword, "STORE", "二号门店店长", 2L);
-        initUser("store3", encodedPassword, "STORE", "三号门店店长", 3L);
+        initUser("admin", encodedPassword, "HQ", "超级管理员", 0);
+        initUser("store1", encodedPassword, "STORE", "一号门店店长", 1);
+        initUser("store2", encodedPassword, "STORE", "二号门店店长", 2);
+        initUser("store3", encodedPassword, "STORE", "三号门店店长", 3);
     }
 
-    private void initUser(String username, String encodedPassword, String role, String realName, Long storeId) {
+    private void initUser(String username, String encodedPassword, String role, String realName, Integer storeId) {
         QueryWrapper<User> wrapper = new QueryWrapper<>();
         wrapper.eq("username", username);
         if (userMapper.selectCount(wrapper) == 0) {
